moved find_mc from clique.cpp and paracl_cp.cpp into its own module: Maxclique_Finder.h,.cpp
